+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## [0.2.1] - 2026-06-27
 ### Changed
+- Change what verbose accepts
 - Improved README: mobile setup instructions, PDF limitation note, Rust install link
 
 ## [0.2.0] - 2026-06-26